Convert reset-test to regular Rust-style tests
